Move `network_filter` to lib folder
